Styling roles on container states + rebasing
