+++ packages/ui/src/memberships/components/MemberRoles.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
-import styled, { css } from 'styled-components'
+import styled from 'styled-components'
 
-import { DefaultTooltip, Tooltip } from '@/common/components/Tooltip'
 import { memberRoleAbbreviation, memberRoleTitle } from '@/memberships/helpers'
 
-import { Colors, Fonts } from '../../common/constants'
+import { MemberRoleHelp, MemberRoleHelpMax, Tooltip } from '../../common/components/Tooltip'
 import { MemberRole } from '../types'
 
 interface MemberRolesProps {
@@ -63,57 +62,11 @@ export const MemberRoles = ({ size, max, wrapable, roles }: MemberRolesProps) =>
   )
 }
 
-interface MemberRoleTooltipProps {
-  size?: 'l' | 'm'
-  isOnDark?: boolean
-}
-
-export const MemberRoleHelp = styled(DefaultTooltip)<MemberRoleTooltipProps>`
-  width: ${({ size }) => (size === 'l' ? '24px' : '16px')};
-  height: ${({ size }) => (size === 'l' ? '24px' : '16px')};
-  font-size: ${({ size }) => (size === 'l' ? '10px' : '6px')};
-  line-height: 1;
-  font-family: ${Fonts.Inter};
-  font-weight: 700;
-  ${({ isOnDark }) =>
-    isOnDark
-      ? css`
-          color: ${Colors.Black[300]};
-          background-color: ${Colors.Black[600]};
-          border-color: ${Colors.Black[600]};
-
-          &:hover,
-          &:focus {
-            color: ${Colors.White};
-            background-color: ${Colors.Blue[500]};
-            border-color: ${Colors.Blue[500]};
-          }
-        `
-      : css`
-          color: ${Colors.Black[600]};
-          background-color: ${Colors.Black[100]};
-          border-color: ${Colors.Black[100]};
-
-          &:hover,
-          &:focus {
-            color: ${Colors.White};
-            background-color: ${Colors.Blue[500]};
-            border-color: ${Colors.Blue[500]};
-          }
-        `};
-`
-
 export const MemberRolesWrapperWrapable = styled.div`
   display: grid;
   grid-template-columns: repeat(auto-fill, minmax(16px, 24px));
   grid-row-gap: 4px;
   grid-column-gap: 4px;
-`
-
-const MemberRoleHelpMax = styled(MemberRoleHelp)`
-  background-color: ${Colors.White};
-  color: ${Colors.Blue[500]};
-  border-color: ${Colors.Blue[50]};
 `
 
 export const MemberRolesWrapper = styled.div`
