+++ packages/ui/src/memberships/components/index.ts
@@ -2,5 +2,4 @@ export * from './components'
 export * from './types'
 export * from './MemberInfo'
 export * from './MembershipCount'
-export { MemberRolesWrapper } from './MemberRoles'
-export { MemberRoleHelp } from './MemberRoles'
+export * from './MemberRoles'
